refactor(mf2): Move MessageValue definition to its own file
