+++ mf2/messageformat/src/resolve/resolve-function-ref.ts
@@ -1,14 +1,14 @@
-import { MessageError } from '../errors.js';
-import type { Context } from '../format-context.js';
-import { fallback } from '../functions/fallback.js';
-import { MessageFunctionContext } from './function-context.js';
-import { getValueSource, resolveValue } from './resolve-value.js';
 import type {
   FunctionRef,
   Literal,
   Options,
   VariableRef
 } from '../data-model/types.js';
+import { MessageError } from '../errors.js';
+import type { Context } from '../format-context.js';
+import { fallback } from '../functions/fallback.js';
+import { MessageFunctionContext } from './function-context.js';
+import { getValueSource, resolveValue } from './resolve-value.js';
 
 export function resolveFunctionRef(
   ctx: Context,
