+++ mf2/messageformat/src/resolve/resolve-variable.ts
@@ -1,11 +1,11 @@
+import type { Expression, VariableRef } from '../data-model/types.js';
 import { MessageResolutionError } from '../errors.js';
 import type { Context } from '../format-context.js';
 import { fallback } from '../functions/fallback.js';
-import { MessageValue } from '../functions/index.js';
 import { unknown } from '../functions/unknown.js';
+import { MessageValue } from '../message-value.js';
 import { MessageFunctionContext } from './function-context.js';
 import { resolveExpression } from './resolve-expression.js';
-import type { Expression, VariableRef } from '../data-model/types.js';
 
 /**
  * Declarations aren't resolved until they're requierd,
